pyproject.toml
# Copyright (c) 2026 Mengzhao Wang
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in
# all copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.
#
# SPDX-License-Identifier: MIT

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hydra-viewer"
version = "0.1.0"
description = "A TUI for Hydra configuration management"
readme = "README.md"
authors = [
    { name = "Mengzhao Wang", email = "mengzhao_wang@foxmail.com" }
]
license = { text = "MIT" }
keywords = ["hydra", "tui", "configuration", "omegaconf", "textual"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "textual[syntax]>=6.00",
    "omegaconf>=2.3",
    "hydra-core>=1.3",
    "pyyaml>=6.0",
    "watchfiles>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/mz-wang/hydra-viewer"
Documentation = "https://github.com/mz-wang/hydra-viewer#readme"
Repository = "https://github.com/mz-wang/hydra-viewer"
Issues = "https://github.com/mz-wang/hydra-viewer/issues"
Changelog = "https://github.com/mz-wang/hydra-viewer/blob/master/CHANGELOG.md"

[project.scripts]
hydra-viewer = "hydra_viewer.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hydra_viewer"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
]
exclude = [
    "*.log",
    "*.pyc",
    "__pycache__",
    ".git",
    ".github",
    ".venv",
    ".python-version",
    "uv.lock",
    ".pytest_cache/",
    "test/",
    "tests/",
    "examples/",
    "docs/",
    "*.egg-info",
    "devmemo/"
]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
    "E",  # All pycodestyle errors
    "F",  # All pyflakes errors
    "I",  # All isort errors
    "UP", # pyupgrade: all pyupgrade errors
    # "S",  # All bandit errors
    "B",  # All flake8-bugbear errors
]
ignore = [
    "E501",  # Line too long (handled by line-length)
    "S101",  # Use of assert detected. The enclosed code will be removed when compiling to optimized bytecode.
]

[tool.ruff.lint.per-file-ignores]
"test/**/*.py" = ["S", "B"]  # Ignore security and bugbear checks in test files

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# note: no tests yet, so commented out for now.
# [tool.pytest.ini_options]
# testpaths = ["tests"]           # test folder
# python_files = ["test_*.py"]
# python_classes = ["Test*"]
# python_functions = ["test_*"]
# addopts = [
#     "-v",
#     "--tb=short",
#     "--cov=src/hydra_viewer",
#     "--cov-report=term-missing",
#     "--cov-report=xml",
# ]

# [tool.coverage.run]
# source = ["src/hydra_viewer"]
# branch = true

# [tool.coverage.report]
# exclude_lines = [
#     "pragma: no cover",
#     "if TYPE_CHECKING:",
#     "raise NotImplementedError",
#     "if __name__ == .__main__.:",
# ]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_scheme = "pep440"
update_changelog_on_bump = true
major_version_zero = true

version_files = [
    "pyproject.toml:^version",
    "src/hydra_viewer/__init__.py:^__version__",
]
